pyproject.toml
[tool.poetry]
name = "hydrocron-wrapper"
version = "0.1.0"
description = "A lightweight Python wrapper for the Hydrocron API"
authors = ["Jake Gearon <jake.gearon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pandas = "^2.2.0"
pydantic = "^2.6.0"
plotly = "^5.24.1"
geopandas = "^1.0.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.10.0"
xarray = "^2025.1.1"
folium = "^0.19.4"
cartopy = "^0.24.1"
sqlalchemy = "^2.0.37"
geoalchemy2 = "^0.17.0"
tqdm = "^4.67.1"
scikit-learn = "^1.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
responses = "^0.24.1"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=hydrocron_wrapper --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
